Setup the project structure for working within the RAP.

Setup the project structure for working within the RAP.

Usage

admin_create_project(path)

Arguments

path

File path and name of project.

Value

Nothing. Used for side effect of making project structure.

Examples

if (FALSE) { # \dontrun{
# Run proj_setup_git_config() first before running the admin_create_project() function.
# ukbAid::proj_setup_git_config("Luke W. Johnston", "lwjohnst@gmail.com")
admin_create_project(fs::path_temp("testproj"))
} # }